duckstation: super close
