+++ Cores/DuckStation/PVDuckStation.xcodeproj/xcshareddata/xcschemes/flac.xcscheme
@@ -15,7 +15,7 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "B3C290C926DE95AE003CB9D4"
-               BuildableName = "libflac.a"
+               BuildableName = "libFLAC.a"
                BlueprintName = "FLAC"
                ReferencedContainer = "container:PVDuckStation.xcodeproj">
             </BuildableReference>
@@ -51,7 +51,7 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "B3C290C926DE95AE003CB9D4"
-            BuildableName = "libflac.a"
+            BuildableName = "libFLAC.a"
             BlueprintName = "FLAC"
             ReferencedContainer = "container:PVDuckStation.xcodeproj">
          </BuildableReference>
