+++ Provenance.xcodeproj/xcshareddata/xcschemes/lzma.xcscheme
@@ -15,7 +15,7 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "B3C291AF26DEED30003CB9D4"
-               BuildableName = "liblzma.a"
+               BuildableName = "libllzma.a"
                BlueprintName = "llzma"
                ReferencedContainer = "container:Cores/DuckStation/PVDuckStation.xcodeproj">
             </BuildableReference>
@@ -51,7 +51,7 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "B3C291AF26DEED30003CB9D4"
-            BuildableName = "liblzma.a"
+            BuildableName = "libllzma.a"
             BlueprintName = "llzma"
             ReferencedContainer = "container:Cores/DuckStation/PVDuckStation.xcodeproj">
          </BuildableReference>
